Add option to enter rate matrix in evolutionary model tab
This commit also refactors some of the code in t2_evolutionarymodel.py,
as well as fixing some bugs with respect to the numeric datatypes being saved.
It should be noted which parameters are floats/ints and when conversions are
necessary or not necessary, and when conversion errors are desireable to produce
error messages.